feat: Move lifecycle_manager to core/managers/
- Move src/fapilog/_internal/lifecycle_manager.py to src/fapilog/core/managers/
- Update import statements in:
  - src/fapilog/_internal/component_factory.py
  - src/fapilog/container.py
  - tests/test_lifecycle_manager.py
  - tests/test_component_factory.py
  - tests/test_container_easy_wins.py
- Fix relative imports within lifecycle_manager.py
- Fix patch paths in test files

✅ LifecycleManager successfully moved to new location (2/12 files moved)
